ref(supergroups): Encapsulate feedback banner in SupergroupFeedback
Move the FeedbackContainer and question text into the
SupergroupFeedback component so the drawer just renders a single
component without needing to know about the feedback layout.
