+++ static/app/views/issueList/supergroups/supergroupFeedback.tsx
@@ -1,4 +1,5 @@
 import {useCallback, useState} from 'react';
+import styled from '@emotion/styled';
 
 import {Button} from '@sentry/scraps/button';
 import {Flex} from '@sentry/scraps/layout';
@@ -32,24 +33,38 @@ export function SupergroupFeedback({supergroupId}: SupergroupFeedbackProps) {
     [supergroupId, organization, user]
   );
 
-  if (feedbackSubmitted) {
-    return <span>{t('Thanks!')}</span>;
-  }
-
   return (
-    <Flex gap="sm">
-      <Button
-        size="zero"
-        icon={<IconThumb direction="up" size="xs" />}
-        onClick={() => handleFeedback(true)}
-        aria-label={t('Yes, this grouping is accurate')}
-      />
-      <Button
-        size="zero"
-        icon={<IconThumb direction="down" size="xs" />}
-        onClick={() => handleFeedback(false)}
-        aria-label={t('No, this grouping is not accurate')}
-      />
-    </Flex>
+    <FeedbackContainer>
+      {feedbackSubmitted ? (
+        t('Thanks!')
+      ) : (
+        <Flex align="center" gap="md">
+          {t('Help us improve this feature. Is this grouping accurate?')}
+          <Flex gap="sm">
+            <Button
+              size="zero"
+              icon={<IconThumb direction="up" size="xs" />}
+              onClick={() => handleFeedback(true)}
+              aria-label={t('Yes, this grouping is accurate')}
+            />
+            <Button
+              size="zero"
+              icon={<IconThumb direction="down" size="xs" />}
+              onClick={() => handleFeedback(false)}
+              aria-label={t('No, this grouping is not accurate')}
+            />
+          </Flex>
+        </Flex>
+      )}
+    </FeedbackContainer>
   );
 }
+
+const FeedbackContainer = styled('div')`
+  display: flex;
+  align-items: center;
+  gap: ${p => p.theme.space.md};
+  padding: ${p => p.theme.space.md} ${p => p.theme.space['2xl']};
+  border-bottom: 1px solid ${p => p.theme.tokens.border.primary};
+  background: ${p => p.theme.tokens.background.transparent.promotion.muted};
+`;
